requirements/xadmin.txt
diff-match-patch==20121119
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
PyYAML==3.12
requirements==0.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
